refactor: solid on fetch Projects and remove components
