+++ src/components/features/index.ts
@@ -1,10 +1,8 @@
 export * from './CardBioSection'
-export * from './CardContainer'
 export * from './CardProfile'
 export * from './CardProjects'
 export * from './DialogConfirmation.component'
 export * from './DialogContact'
-export * from './DialogContainer'
 export * from './StatusErrorMessage'
 export * from './StatusLoadingMessage'
 export * from './StatusWarnMessage'
