+++ src/hooks/fetch/fetchProject.hook.ts
@@ -2,7 +2,7 @@
 
 import { trpc } from '@/utils/trpc/trpc'
 
-export function GetProjects() {
+export function fetchProjects() {
   return {
     getAllProjects: trpc.project.getAllProjects.useQuery(),
   }
